backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-nmap==0.7.1
ollama==0.4.5
reportlab==4.4.5
Pillow==12.0.0
pydantic==2.10.4
requests==2.31.0
